refactor : exists 메서드 추가
객체를 불러와서 null 판단하는 것 보다는, 로직을 한번에 처리하는 것이 좋지 않을까 생각

+++ src/main/java/com/aiary/be/user/application/UserService.java
@@ -35,13 +35,13 @@ public class UserService {
 
         user.update(request.email(), request.password(), request.userName(), request.age(),
             request.gender(), request.phoneNumber(), passwordEncoder);
-
     }
 
     public void deleteUser(Long userId){
-        userRepository.findById(userId)
-            .orElseThrow(() -> CustomException.from(UserErrorCode.NOT_FOUND));
-
+        if(!userRepository.existsById(userId)) {
+            throw CustomException.from(UserErrorCode.NOT_FOUND);
+        }
+        
         userRepository.deleteById(userId);
     }
